requirements.txt
pymongo[tls,srv]==4.0.1
python-dotenv==0.19.2
Flask==2.0.2
flask-swagger==0.2.14
requests
dnspython==2.0.0